feat:change profile implementation, devide profile on user information and his security and add some features to this 2 components
