+++ src/app/profile/models/profile.model.ts
@@ -1,6 +1,7 @@
 export interface IProfile{
     first_name: string,
     last_name: string,
+    email: string,
     role: string,
     phone:string
 }
