+++ src/app/profile/pages/profile-details/profile-details.component.html
@@ -1,0 +1,95 @@
+<div class="profile">
+    <h1>Profile details</h1>
+    <div class="role" *ngIf="(profile$|async)?.role=='admin'">
+        Manager
+    </div>
+    <div class="profile-details-line"></div>
+
+    <form  class="form-block" [formGroup]="profileForm" >
+
+        <div class="profile-data" *ngIf="!isEditingFirstName">
+            <h2>First name:</h2> 
+            <p>{{ profileForm.controls['firstName'].value }}  <button class="profile-change-button" (click)="editFirstName()"><mat-icon>edit</mat-icon></button></p>
+        </div>
+
+        <div *ngIf="isEditingFirstName">
+            <div class="input-block">
+                <label for="first-name">First name:</label>
+                <div>
+                    <input id="first-name" type="text" formControlName="firstName" >
+                    <button class="profile-change-button" (click)="saveFirstName()" [disabled]="profileForm.controls['firstName'].invalid"><mat-icon>save</mat-icon></button>
+                </div>
+                @if(profileForm.controls['firstName'].invalid){
+                    <div class="alert">
+                        Required
+                    </div>
+                }
+            </div>
+        </div>
+        <div class="profile-details-line"></div>
+
+        <div class="profile-data" *ngIf="!isEditingLastName">
+            <h2>Last name:</h2> 
+            <p>{{ profileForm.controls['lastName'].value }}  <button class="profile-change-button" (click)="editLastName()"><mat-icon>edit</mat-icon></button></p>
+        </div>
+
+        <div *ngIf="isEditingLastName">
+            <div class="input-block">
+                <label for="last-name">Last name:</label>
+                <div>
+                    <input id="last-name" type="text" formControlName="lastName" >
+                    <button class="profile-change-button" (click)="saveLastName()" [disabled]="profileForm.controls['lastName'].invalid"><mat-icon>save</mat-icon></button>
+                </div>
+                @if(profileForm.controls['lastName'].invalid){
+                    <div class="alert">
+                        Required
+                    </div>
+                }
+            </div>
+        </div>
+        <div class="profile-details-line"></div>
+
+        <div class="profile-data" *ngIf="!isEditingPhone">
+            <h2>Phone:</h2> 
+            <p>{{ profileForm.controls['phone'].value }}  <button class="profile-change-button" (click)="editPhone()"><mat-icon>edit</mat-icon></button></p>
+        </div>
+
+        <div *ngIf="isEditingPhone">
+            <div class="input-block">
+                <label for="phone">Phone:</label>
+                <div>
+                    <input id="phone" formControlName="phone" [mask]="'0 (000) 000-0000'" >
+                    <button class="profile-change-button" (click)="savePhone()" [disabled]="profileForm.controls['phone'].invalid"><mat-icon>save</mat-icon></button>
+                </div>
+                @if(profileForm.controls['phone'].invalid){
+                    <div class="alert">
+                        Required
+                    </div>
+                }
+            </div>
+        </div>
+        <div class="profile-details-line"></div>
+
+        <div class="profile-data" *ngIf="!isEditingEmail">
+            <h2>Email:</h2> 
+            <p>{{ profileForm.controls['email'].value }}  <button class="profile-change-button" (click)="editEmail()"><mat-icon>edit</mat-icon></button></p>
+        </div>
+
+        <div *ngIf="isEditingEmail">
+            <div class="input-block">
+                <label for="email">Email:</label>
+                <div>
+                    <input id="email" formControlName="email">
+                    <button class="profile-change-button" (click)="saveEmail()" [disabled]="profileForm.controls['email'].invalid"><mat-icon>save</mat-icon></button>
+                </div>
+                @if(profileForm.controls['email'].invalid){
+                    <div class="alert">
+                        Required
+                    </div>
+                }
+            </div>
+        </div>
+    </form>
+
+    
+</div>
